+++ engine/glitch/core/uid.h
@@ -2,26 +2,29 @@
 
 namespace gl {
 
+// TODO: we made this 32-bits for lua compability but if you ever want
+// more concurrent entities (4.3B is the limit now) make this 64-bit again
+
 /**
- * 64-bit randomized unique identifier.
+ * 32-bit randomized unique identifier.
  */
 struct GL_API UID {
-	uint64_t value;
+	uint32_t value;
 
 	UID();
-	UID(const uint64_t& p_uuid);
-	UID(uint64_t&& p_uuid);
+	UID(const uint32_t& p_uuid);
+	UID(uint32_t&& p_uuid);
 	UID(const UID&) = default;
 
 	UID& operator=(const UID& p_other);
 	UID& operator=(UID&& p_other);
 
-	UID& operator=(const uint64_t& p_other);
-	UID& operator=(uint64_t&& p_other);
+	UID& operator=(const uint32_t& p_other);
+	UID& operator=(uint32_t&& p_other);
 
 	bool is_valid() const { return value != 0; }
 
-	operator uint64_t() const { return value; }
+	operator uint32_t() const { return value; }
 };
 
 inline const UID INVALID_UID = 0;
@@ -32,6 +35,6 @@ namespace std {
 template <typename T> struct hash;
 
 template <> struct hash<gl::UID> {
-	size_t operator()(const gl::UID& p_uuid) const { return (uint64_t)p_uuid; }
+	size_t operator()(const gl::UID& p_uuid) const { return (uint32_t)p_uuid; }
 };
 } //namespace std
